feat: Modify python flask files to MVC structure and Modify image caresoull in attraction
