+++ models/booking.py
@@ -1,0 +1,9 @@
+class Booking:
+    def get():
+        pass
+
+    def post():
+        pass
+
+    def delete():
+        pass
