+++ models/category.py
@@ -1,0 +1,4 @@
+class Category:
+    def get(cursor):
+        cursor.execute("select category from attraction")
+        return cursor.fetchall()
