+++ models/order.py
@@ -1,0 +1,6 @@
+class Order:
+    def get():
+        pass
+
+    def post():
+        pass
