+++ static/css/attraction.css
@@ -1,5 +1,14 @@
 @import "../page.css";
 
+@keyframes fadein {
+  0% {
+    opacity: 0;
+  }
+  100% {
+    opacity: 1;
+  }
+}
+
 .navigation-bar-container {
   border-bottom: 1px solid #e8e8e8;
 }
@@ -22,18 +31,26 @@
   padding-bottom: 37px;
 }
 
-.image {
-  height: 406px;
-  display: flex;
-  justify-content: center;
+.images-container {
   position: relative;
   border-radius: 5px;
 }
 
-#attractionImage {
-  width: 100%;
-  height: 100%;
+.sliders {
+  height: 406px;
+  display: flex;
+  overflow: hidden;
   border-radius: 5px;
+  transition: all 0.5s ease;
+}
+
+.fade-in-and-out {
+  animation: fadein 1s ease-in-out;
+}
+
+.attraction-image {
+  width: 540px;
+  height: 100%;
 }
 
 .arrow {
@@ -55,8 +72,9 @@
   right: 10px;
 }
 
-#imageOrder {
+#paginationDots {
   position: absolute;
+  width: 100%;
   bottom: 6px;
   display: flex;
   justify-content: center;
@@ -64,7 +82,7 @@
   cursor: pointer;
 }
 
-#imageOrder > div div {
+#paginationDots > div div {
   min-width: 10px;
   min-height: 10px;
   border: 2px solid #ffffff;
@@ -190,13 +208,14 @@
     border-radius: 0;
   }
 
-  .image {
-    width: 100vw;
+  .sliders {
+    width: 100%;
     height: calc((400 / 540) * 100vw);
+    border-radius: 0;
   }
 
-  #attractionImage {
-    border-radius: 0;
+  .attraction-image {
+    width: 100vw;
   }
 
   .attraction-info {
