+++ static/js/attraction.js
@@ -1,6 +1,6 @@
 let url = "http://0.0.0.0:3000/";
-let attractionImage = document.querySelector("#attractionImage");
-let imageOrder = document.querySelector("#imageOrder");
+let sliders = document.querySelector(".sliders");
+let paginationDots = document.querySelector("#paginationDots");
 let currentImage = 0;
 let maxImage;
 let imageList = [];
@@ -20,18 +20,18 @@ const loadAttractionById = () => {
   let attractionLocation = window.location.href;
   let attractionIdIndex = attractionLocation.lastIndexOf("/");
   let attractionId = attractionLocation.substring(attractionIdIndex + 1);
-
+  
   let attractionName = document.querySelector("#attractionName");
   let attractionCtegoryAndMrt = document.querySelector(
     "#attractionCtegoryAndMrt"
-  );
-  let description = document.querySelector("#description");
-  let address = document.querySelector("#address");
-  let transport = document.querySelector("#transport");
-
-  morning.click();
-  apiUrl += attractionId;
-  fetch(apiUrl)
+    );
+    let description = document.querySelector("#description");
+    let address = document.querySelector("#address");
+    let transport = document.querySelector("#transport");
+    
+    morning.click();
+    apiUrl += attractionId;
+    fetch(apiUrl)
     .then((response) => {
       return response.json();
     })
@@ -43,42 +43,55 @@ const loadAttractionById = () => {
         description,
         address,
         transport
-      );
-
-      changeImageOrder(currentImage);
-    });
-};
-
-const loadData = (jsonResult, ...doms) => {
-  imageList = jsonResult["data"]["images"];
-  attractionImage.src = imageList[0];
-  maxImage = imageList.length - 1;
-  for (let i = 0; i < imageList.length; i++) {
-    let orderDotContainer = document.createElement("div");
-    let orderDot = document.createElement("div");
-    orderDotContainer.id = "image" + i;
-    orderDot.id = "image" + i;
-    orderDotContainer.appendChild(orderDot);
-    imageOrder.appendChild(orderDotContainer);
-  }
-
-  doms[0].textContent = jsonResult["data"]["name"];
-  doms[1].textContent =
-    jsonResult["data"]["category"] + " at " + jsonResult["data"]["mrt"];
-  doms[2].textContent = jsonResult["data"]["description"];
-  doms[3].textContent = jsonResult["data"]["address"];
-  doms[4].textContent = jsonResult["data"]["transport"];
-};
-
-const changeImage = (e) => {
-  let imageNumber = parseInt(e.target.id.replace("image", ""));
-  if (e.target.id === "leftArrow") {
-    if (currentImage === 0) {
-      currentImage = maxImage;
-    } else {
-      currentImage--;
-    }
-  } else if (e.target.id === "rightArrow") {
+        );
+        changePaginationDot(currentImage);
+      });
+    };
+    
+    const loadData = (jsonResult, ...doms) => {
+      imageList = jsonResult["data"]["images"];
+      loadImage();
+      maxImage = imageList.length - 1;
+      for (let i = 0; i < imageList.length; i++) {
+        let paginationDotContainer = document.createElement("div");
+        let paginationDot = document.createElement("div");
+        paginationDotContainer.id = "image" + i;
+        paginationDot.id = "image" + i;
+        paginationDotContainer.appendChild(paginationDot);
+        paginationDots.appendChild(paginationDotContainer);
+      }
+      
+      doms[0].textContent = jsonResult["data"]["name"];
+      doms[1].textContent =
+      jsonResult["data"]["category"] + " at " + jsonResult["data"]["mrt"];
+      doms[2].textContent = jsonResult["data"]["description"];
+      doms[3].textContent = jsonResult["data"]["address"];
+      doms[4].textContent = jsonResult["data"]["transport"];
+    };
+    
+    const loadImage = () => {
+      imageList.forEach((imageSrc) => {
+        let attractionImage = document.createElement("img");
+        attractionImage.src = imageSrc;
+        attractionImage.classList.add("attraction-image");
+        let imageContainer = document.createElement("div");
+        imageContainer.appendChild(attractionImage);
+        sliders.appendChild(imageContainer);
+        sliders.classList.add("fade-in-and-out");
+      });
+    };
+    
+    const changeImage = (e) => {
+      let slideWidth = sliders.offsetWidth;
+      let imageNumber = parseInt(e.target.id.replace("image", ""));
+      sliders.classList.remove("fade-in-and-out");
+      if (e.target.id === "leftArrow") {
+        if (currentImage === 0) {
+          currentImage = maxImage;
+        } else {
+          currentImage--;
+        }
+      } else if (e.target.id === "rightArrow") {
     if (currentImage === maxImage) {
       currentImage = 0;
     } else {
@@ -87,16 +100,17 @@ const changeImage = (e) => {
   } else {
     currentImage = imageNumber;
   }
-  attractionImage.src = imageList[currentImage];
-  changeImageOrder(currentImage);
+  sliders.scrollLeft = currentImage * slideWidth;
+  sliders.classList.add("fade-in-and-out");
+  changePaginationDot(currentImage);
 };
 
-const changeImageOrder = (currentImg) => {
-  let imageOrderChilds = imageOrder.childNodes;
-  imageOrderChilds.forEach((element) => {
+const changePaginationDot = (currentImg) => {
+  let paginationDotChilds = paginationDots.childNodes;
+  paginationDotChilds.forEach((element) => {
     element.childNodes[0].style.backgroundColor = "#ffffff";
   });
-  imageOrderChilds[currentImage].childNodes[0].style.backgroundColor =
+  paginationDotChilds[currentImage].childNodes[0].style.backgroundColor =
     "#000000";
 };
 
@@ -123,4 +137,4 @@ arrowList.forEach((element) => {
   element.addEventListener("click", changeImage, false);
 });
 
-imageOrder.addEventListener("click", changeImage, false);
+paginationDots.addEventListener("click", changeImage, false);
